mgr: cut down duplication between active+standby
...by using PyModuleRunner class from ActivePyModule too.

(cherry picked from commit df8797320bed7ad9f121477e35d7e3862efd89bd)
