+++ src/mgr/StandbyPyModules.cc
@@ -33,13 +33,6 @@
 // Declaration fulfilled by ActivePyModules
 std::string handle_pyerror();
 
-void* PyModuleRunner::PyModuleRunnerThread::entry()
-{
-  // No need to acquire the GIL here; the module does it.
-  dout(4) << "Entering thread for " << mod->get_name() << dendl;
-  mod->serve();
-  return nullptr;
-}
 
 StandbyPyModules::StandbyPyModules(MonClient *monc_, const MgrMap &mgr_map_)
     : monc(monc_), load_config_thread(monc, &state)
@@ -138,19 +131,6 @@ int StandbyPyModule::load()
   }
 }
 
-StandbyPyModule::~StandbyPyModule()
-{
-  Gil gil(pMyThreadState);
-
-  if (pClassInstance) {
-    Py_XDECREF(pClassInstance);
-    pClassInstance = nullptr;
-  }
-
-  Py_DECREF(pClass);
-  pClass = nullptr;
-}
-
 void *StandbyPyModules::LoadConfigThread::entry()
 {
   dout(10) << "listing keys" << dendl;
@@ -215,54 +195,5 @@ std::string StandbyPyModule::get_active_uri() const
   });
 
   return result;
-}
-
-int PyModuleRunner::serve()
-{
-  assert(pClassInstance != nullptr);
-
-  // This method is called from a separate OS thread (i.e. a thread not
-  // created by Python), so tell Gil to wrap this in a new thread state.
-  Gil gil(pMyThreadState, true);
-
-  auto pValue = PyObject_CallMethod(pClassInstance,
-      const_cast<char*>("serve"), nullptr);
-
-  int r = 0;
-  if (pValue != NULL) {
-    Py_DECREF(pValue);
-  } else {
-    derr << module_name << ".serve:" << dendl;
-    derr << handle_pyerror() << dendl;
-    return -EINVAL;
-  }
-
-  return r;
-}
-
-void PyModuleRunner::shutdown()
-{
-  assert(pClassInstance != nullptr);
-
-  Gil gil(pMyThreadState);
-
-  auto pValue = PyObject_CallMethod(pClassInstance,
-      const_cast<char*>("shutdown"), nullptr);
-
-  if (pValue != NULL) {
-    Py_DECREF(pValue);
-  } else {
-    derr << "Failed to invoke shutdown() on " << module_name << dendl;
-    derr << handle_pyerror() << dendl;
-  }
-}
-
-void PyModuleRunner::log(int level, const std::string &record)
-{
-#undef dout_prefix
-#define dout_prefix *_dout << "mgr[" << module_name << "] "
-  dout(level) << record << dendl;
-#undef dout_prefix
-#define dout_prefix *_dout << "mgr " << __func__ << " "
 }
